MNT: Move tools out of application and main window.
